+++ src/Adapter/WorldlineSDKAdapter.php
@@ -149,22 +149,6 @@ class WorldlineSDKAdapter
     }
 
     /**
-     * @return void
-     * @throws \Exception
-     */
-    public function testConnection()
-    {
-        $queryParams = new GetPaymentProductsParams();
-
-        $queryParams->setCountryCode("DEU");
-        $queryParams->setCurrencyCode("EUR");
-
-        $this->merchantClient
-            ->products()
-            ->getPaymentProducts($queryParams);
-    }
-
-    /**
      * @param string $hostedCheckoutId
      * @return PaymentDetailsResponse
      * @throws \Exception
@@ -181,14 +165,16 @@ class WorldlineSDKAdapter
      * @param string $currencyISO
      * @param int $worldlinePaymentProductId
      * @param OrderEntity|null $orderEntity
+     * @param string $token
      * @return CreateHostedCheckoutResponse
-     * @throws \Exception
+     * @throws Exception
      */
     public function createPayment(
         int          $amountTotal,
         string       $currencyISO,
         int          $worldlinePaymentProductId,
-        ?OrderEntity $orderEntity
+        ?OrderEntity $orderEntity,
+        string       $token
     ): CreateHostedCheckoutResponse
     {
         $fullRedirectTemplateName = $this->getPluginConfig(Form::FULL_REDIRECT_TEMPLATE_NAME);
@@ -237,10 +223,15 @@ class WorldlineSDKAdapter
             );
         }
 
+        if ($token != '') {
+            $cardPaymentMethodSpecificInput->setToken($token);
+        }
+
         $hostedCheckoutRequest->setOrder($order);
         $hostedCheckoutRequest->setHostedCheckoutSpecificInput($hostedCheckoutSpecificInput);
         $hostedCheckoutRequest->setCardPaymentMethodSpecificInput($cardPaymentMethodSpecificInput);
         $hostedCheckoutClient = $merchantClient->hostedCheckout();
+
         return $hostedCheckoutClient->createHostedCheckout($hostedCheckoutRequest);
     }
 
@@ -280,7 +271,6 @@ class WorldlineSDKAdapter
                 );
                 $redirectPaymentMethodSpecificInput = new RedirectPaymentMethodSpecificInput();
                 $redirectPaymentMethodSpecificInput->setPaymentProductId($worldlinePaymentProductId);
-                $hostedCheckoutRequest->setRedirectPaymentMethodSpecificInput($redirectPaymentMethodSpecificInput);
                 break;
             }
             case PaymentProducts::PAYMENT_PRODUCT_ONEY_3X_4X:
@@ -294,9 +284,12 @@ class WorldlineSDKAdapter
                 $redirectPaymentMethodSpecificInput->setPaymentProductId($worldlinePaymentProductId);
                 $redirectPaymentMethodSpecificInput->setRequiresApproval(true);
                 $redirectPaymentMethodSpecificInput->setPaymentOption($this->getPluginConfig(Form::ONEY_PAYMENT_OPTION_FIELD));
-                $hostedCheckoutRequest->setRedirectPaymentMethodSpecificInput($redirectPaymentMethodSpecificInput);
                 break;
             }
+        }
+
+        if (isset($redirectPaymentMethodSpecificInput)) {
+            $hostedCheckoutRequest->setRedirectPaymentMethodSpecificInput($redirectPaymentMethodSpecificInput);
         }
     }
 
@@ -616,10 +609,10 @@ class WorldlineSDKAdapter
      */
     public function getReturnUrl(): string
     {
-        if (!$this->isLiveMode()) {
-            return $this->getPluginConfig(Form::RETURN_URL_FIELD);
+        if ($this->isLiveMode()) {
+            return $this->getPluginConfig(Form::LIVE_RETURN_URL_FIELD);
         }
-        return $this->getPluginConfig(Form::LIVE_RETURN_URL_FIELD);
+        return $this->getPluginConfig(Form::RETURN_URL_FIELD);
     }
 
     /**