+++ src/Controller/TransactionsControl/TransactionsControlController.php
@@ -19,10 +19,8 @@ use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionEnti
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionStateHandler;
 use Shopware\Core\Checkout\Order\OrderEntity;
 use Shopware\Core\Framework\Context;
-use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
+use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
-use Shopware\Core\Framework\Routing\Annotation\RouteScope;
-use Shopware\Core\Kernel;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -32,13 +30,13 @@ use Shopware\Core\System\SystemConfig\SystemConfigService;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
- * @RouteScope(scopes={"api"})
+ * @Route(defaults={"_routeScope"={"api"}})
  */
 class TransactionsControlController extends AbstractController
 {
     private SystemConfigService $systemConfigService;
-    private EntityRepositoryInterface $orderRepository;
-    private EntityRepositoryInterface $customerRepository;
+    private EntityRepository $orderRepository;
+    private EntityRepository $customerRepository;
     private OrderTransactionStateHandler $transactionStateHandler;
     private Logger $logger;
     private TranslatorInterface $translator;
@@ -46,8 +44,8 @@ class TransactionsControlController extends AbstractController
 
     /**
      * @param SystemConfigService $systemConfigService
-     * @param EntityRepositoryInterface $orderRepository
-     * @param EntityRepositoryInterface $customerRepository
+     * @param EntityRepository $orderRepository
+     * @param EntityRepository $customerRepository
      * @param OrderTransactionStateHandler $transactionStateHandler
      * @param Logger $logger
      * @param TranslatorInterface $translator
@@ -55,8 +53,8 @@ class TransactionsControlController extends AbstractController
      */
     public function __construct(
         SystemConfigService          $systemConfigService,
-        EntityRepositoryInterface    $orderRepository,
-        EntityRepositoryInterface    $customerRepository,
+        EntityRepository             $orderRepository,
+        EntityRepository             $customerRepository,
         OrderTransactionStateHandler $transactionStateHandler,
         Logger                       $logger,
         TranslatorInterface          $translator,
@@ -152,7 +150,7 @@ class TransactionsControlController extends AbstractController
         $transaction = $orderEntity->getTransactions()->last();
         $customFields = $transaction->getPaymentMethod()->getCustomFields();
         $isFullRedirectMethod = false;
-        if (array_key_exists(Form::CUSTOM_FIELD_WORLDLINE_PAYMENT_METHOD_ID, $customFields)) {
+        if (is_array($customFields) && array_key_exists(Form::CUSTOM_FIELD_WORLDLINE_PAYMENT_METHOD_ID, $customFields)) {
             $isFullRedirectMethod = $customFields[Form::CUSTOM_FIELD_WORLDLINE_PAYMENT_METHOD_ID] == Payment::FULL_REDIRECT_PAYMENT_METHOD_ID;
         }
 
@@ -264,7 +262,7 @@ class TransactionsControlController extends AbstractController
     private function processPayment(Request $request, Context $context, string $action): JsonResponse
     {
         $hostedCheckoutId = $request->request->get('transactionId');
-        $itemsChanges = $request->request->get('items');
+        $itemsChanges = $request->request->all('items');
 
         $amount = (int)round($request->request->get('amount') * 100);
         if (!$hostedCheckoutId) {