+++ src/Resources/app/administration/src/component/payment-method-button/index.js
@@ -16,6 +16,8 @@ Component.register('payment-method-button', {
         return {
             isLoading: false,
             isSaveSuccessful: false,
+            paymentMethodData: [],
+            displayMessage: null,
         };
     },
 
@@ -43,29 +45,33 @@ Component.register('payment-method-button', {
             return  $parent.currentSalesChannelId;
         },
 
-        selectCheckbox() {
-            document.querySelectorAll('.paymentMethod').forEach((el, number,parent) => {
-                if (number === 0) {
-                    el.click();
-                } else if (el.checked != parent[0].checked) {
-                    el.click();
-                }
-            })
+        selectAllCheckbox(shouldBeActive) {
+            this.paymentMethodData.forEach(item => {
+                item.isActive = shouldBeActive;
+            });
+            this.saveButton();
+        },
+
+        selectSingleCheckbox(internalId) {
+            const element = this.paymentMethodData.find(value => value.internalId === internalId);
+            element.isActive = !element.isActive;
+            this.saveButton();
         },
 
         createPaymentMethodsArray() {
             let paymentMethodsArray = [];
-            document.querySelectorAll('.payment-method--container').forEach((el) => {
+            this.paymentMethodData.forEach(item => {
                 paymentMethodsArray.push({
-                    id: el.children[0].children[0].id,
-                    status: el.children[0].children[0].checked,
-                    internalId: el.children[0].children[0].getAttribute('internalId'),
+                    id: item.id,
+                    status: item.isActive,
+                    internalId: item.internalId,
                 });
-            })
+            });
             return paymentMethodsArray;
         },
 
         saveButton() {
+            this.isLoading = true;
             this.apiTest.savemethod(
             {
                 'data': this.createPaymentMethodsArray(),
@@ -74,10 +80,6 @@ Component.register('payment-method-button', {
             ).then((res) => {
                 if (res.success) {
                     this.isSaveSuccessful = true;
-                    this.createNotificationSuccess({
-                        title: this.$tc('worldline.payment-method-button.APITitle'),
-                        message: this.$tc('worldline.payment-method-button.success')
-                    });
                 } else {
                     this.createNotificationError({
                         title: this.$tc('worldline.payment-method-button.APITitle'),
@@ -90,41 +92,32 @@ Component.register('payment-method-button', {
 
         },
 
-        display(content) {
-            document.querySelector('.border-inner').innerHTML = content;
-        },
-
         renderPaymentMethods(paymentMethods) {
-            if (paymentMethods.length <= 5) {
-                document.querySelector('.select-all').innerHTML = '';
-            }
-            if (paymentMethods.length == 0) {
-                this.display(`<p class="innerText">${this.$tc('worldline.payment-method-button.requestEmpty')}</p>`);
+            if (paymentMethods.length === 0) {
+                this.displayMessage = this.$tc('worldline.payment-method-button.requestEmpty');
             } else {
-                let checkboxTemplate = paymentMethods.map((item) => {
-
-                    return `<div class="payment-method--container">
-                    <label class="switch">
-                      <input type="checkbox" id="${item.id}" internalId="${item.internalId}" class="paymentMethod" ${item.isActive?'checked':''}>
-                      <span class="slider round"></span>
-                    </label>
-                    <img src="${item.logo}">
-                    <label for="${item.label}">
-                    ${item.label}</label>
-                    </div>`;
-                }).join(" ");
-                this.display(checkboxTemplate);
+                this.displayMessage = null;
+                this.paymentMethodData = [];
+                paymentMethods.forEach((item) => {
+                    this.paymentMethodData.push({
+                        id: item.id,
+                        internalId: item.internalId,
+                        isActive: item.isActive,
+                        logo: item.logo,
+                        label: item.label,
+                    });
+                });
             }
         },
 
         getPaymentMethods() {
-            this.display(`<p class="innerText">${this.$tc('worldline.payment-method-button.request')}</p>`);
+            this.displayMessage = this.$tc('worldline.payment-method-button.request');
             this.isLoading = true;
             this.apiTest.check(this.pluginConfig).then((res) => {//todo  split this
                 if (res.success) {
                     this.renderPaymentMethods(res.paymentMethods);
                 } else {
-                    this.display(`<p class="innerText">${this.$tc('worldline.payment-method-button.error')}</p>`);
+                    this.displayMessage = this.$tc('worldline.payment-method-button.error');
                     this.createNotificationError({
                         title: this.$tc('worldline.payment-method-button.title'),
                         message: this.$tc('worldline.payment-method-button.error') + res.message
@@ -132,7 +125,8 @@ Component.register('payment-method-button', {
                 }
                 this.isLoading = false;
             }).catch((error) => {
-                this.display(`<p class="innerText">${this.$tc('worldline.payment-method-button.error')}</p>`);
+                this.displayMessage = this.$tc('worldline.payment-method-button.error');
+                this.isLoading = false;
             });
         }
     },